+++ board/sw/Io/const.hpp
@@ -2,6 +2,6 @@
 
 namespace Io
 {
-constexpr uint8_t max_line_size    = 32;
-constexpr uint8_t max_buffer_size = 2 * max_line_size;
+constexpr size_t max_line_size   = 32;
+constexpr size_t max_buffer_size = 2 * max_line_size;
 }
